+++ apps/client/src/app/model/common/zone-breakdown.ts
@@ -2,17 +2,18 @@ import { ZoneBreakdownRow } from './zone-breakdown-row';
 import { ListRow } from '../../modules/list/model/list-row';
 import { tpWindowEntries } from '../../core/data/sources/tp-window-entries';
 import { LayoutRowFilter } from '../../core/layout/layout-row-filter';
-import { DataType, getItemSource } from '@ffxiv-teamcraft/types';
+import { DataType, getItemSource, Vector2 } from '@ffxiv-teamcraft/types';
 import { Drop } from '../../modules/list/model/drop';
 import { PersistedAlarm } from '../../core/alarms/persisted-alarm';
 import { Vendor } from '../../modules/list/model/vendor';
 import { TradeSource } from '../../modules/list/model/trade-source';
-import { Vector2 } from '@ffxiv-teamcraft/types';
 import { mapIds } from '../../core/data/sources/map-ids';
+import { StaticData } from '../../lazy-data/static-data';
+import { SettingsService } from '../../modules/settings/settings.service';
 
 export class ZoneBreakdown {
 
-  constructor(rows: ListRow[], filterChain?: string, hideZoneDuplicates = false, private finalItems = false) {
+  constructor(rows: ListRow[], private settings: SettingsService, filterChain?: string, hideZoneDuplicates = false, private finalItems = false) {
     rows.forEach(row => {
       if (getItemSource(row, DataType.GATHERED_BY, true).nodes !== undefined && getItemSource(row, DataType.GATHERED_BY, true).nodes.length !== 0
         && this.hasOneFilter(filterChain, LayoutRowFilter.IS_GATHERING, LayoutRowFilter.IS_GATHERED_BY_BTN, LayoutRowFilter.IS_GATHERED_BY_MIN, LayoutRowFilter.IS_GATHERED_BY_FSH)) {
@@ -101,10 +102,24 @@ export class ZoneBreakdown {
   private addToBreakdown(zoneId: number, mapId: number, item: ListRow, hideZoneDuplicates: boolean, coords: Vector2, fateId?: number): void {
     const existingRow = this.rows.find(r => r.mapId === mapId);
     // If we hide duplicates and it's bicolor gems, ignore "all items" shop that needs to be unlocked
-    if (hideZoneDuplicates
-      && getItemSource(item, DataType.TRADE_SOURCES).some(ts => ts.trades.some(t => t.currencies.some(c => c.id === 26807)))
-      && [497, 498, 693, 694, 856].includes(mapId)) {
-      return;
+    const gemstoneTrades = getItemSource(item, DataType.TRADE_SOURCES).filter(ts => ts.trades.some(t => t.currencies.some(c => c.id === 26807)));
+    const isBicolorTrade = hideZoneDuplicates && gemstoneTrades;
+    if (isBicolorTrade) {
+      if (StaticData.globalFATEShopMapIds.includes(mapId)) {
+        // If user unlocked this FATE trade NPC, prioritize it, else skip it
+        if (this.settings.unlockedFATEAreas.includes(mapId)) {
+          this.removeRowsForItem(item.id);
+        } else {
+          return;
+        }
+      } else {
+        // Else, it's a bicolor gemstone shop but not main city
+
+        // If user unlocked global FATE shop for this trade, ignore the current row because we'll want to prioritize global anyways
+        if (gemstoneTrades.some(ts => ts.npcs.some(npc => this.settings.unlockedFATEAreas.includes(npc.mapId)))) {
+          return;
+        }
+      }
     }
     if (hideZoneDuplicates) {
       if (this.rows.some(r => r.items.some(i => i.id === item.id && item.finalItem === i.finalItem))) {