+++ apps/client/src/app/modules/list/list-details-panel/list-details-panel.component.ts
@@ -1,10 +1,10 @@
-import { ChangeDetectionStrategy, ChangeDetectorRef, Component, Input, OnChanges, OnInit, SimpleChanges } from '@angular/core';
+import { ChangeDetectionStrategy, ChangeDetectorRef, Component, Input, OnChanges, OnInit } from '@angular/core';
 import { LayoutRowDisplay } from '../../../core/layout/layout-row-display';
 import { ListRow } from '../model/list-row';
 import { ZoneBreakdownRow } from '../../../model/common/zone-breakdown-row';
 import { I18nToolsService } from '../../../core/tools/i18n-tools.service';
 import { NzModalService } from 'ng-zorro-antd/modal';
-import { TranslateService, TranslateModule } from '@ngx-translate/core';
+import { TranslateModule, TranslateService } from '@ngx-translate/core';
 import { ZoneBreakdown } from '../../../model/common/zone-breakdown';
 import { TotalPanelPricePopupComponent } from '../../../pages/list-details/total-panel-price-popup/total-panel-price-popup.component';
 import { NavigationMapComponent } from '../../map/navigation-map/navigation-map.component';
@@ -26,7 +26,6 @@ import { TradeSource } from '../model/trade-source';
 import { Vendor } from '../model/vendor';
 import { LayoutRowDisplayMode } from '../../../core/layout/layout-row-display-mode';
 import { NpcBreakdown } from '../../../model/common/npc-breakdown';
-import { NpcBreakdownRow } from '../../../model/common/npc-breakdown-row';
 import { LazyDataFacade } from '../../../lazy-data/+state/lazy-data.facade';
 import { safeCombineLatest } from '../../../core/rxjs/safe-combine-latest';
 import { observeInput } from '../../../core/rxjs/observe-input';
@@ -53,15 +52,15 @@ import { AggregateItemRowComponent } from '../item/aggregate-item-row/aggregate-
 import { ItemRowComponent } from '../item/item-row/item-row.component';
 import { LazyScrollComponent } from '../../lazy-scroll/lazy-scroll/lazy-scroll.component';
 import { NzCollapseModule } from 'ng-zorro-antd/collapse';
-import { NgTemplateOutlet, AsyncPipe, DecimalPipe } from '@angular/common';
+import { AsyncPipe, DecimalPipe, NgTemplateOutlet } from '@angular/common';
 
 @Component({
-    selector: 'app-list-details-panel',
-    templateUrl: './list-details-panel.component.html',
-    styleUrls: ['./list-details-panel.component.less'],
-    changeDetection: ChangeDetectionStrategy.OnPush,
-    standalone: true,
-    imports: [NzCollapseModule, LazyScrollComponent, ItemRowComponent, AggregateItemRowComponent, NzGridModule, NgForTrackByIdDirective, CompactItemRowComponent, NzDividerModule, NgTemplateOutlet, FlexModule, NzButtonModule, NzIconModule, NzToolTipModule, NzWaveModule, MapPositionComponent, NzProgressModule, NzPopconfirmModule, ClipboardDirective, NzSpinModule, AsyncPipe, DecimalPipe, I18nPipe, TranslateModule]
+  selector: 'app-list-details-panel',
+  templateUrl: './list-details-panel.component.html',
+  styleUrls: ['./list-details-panel.component.less'],
+  changeDetection: ChangeDetectionStrategy.OnPush,
+  standalone: true,
+  imports: [NzCollapseModule, LazyScrollComponent, ItemRowComponent, AggregateItemRowComponent, NzGridModule, NgForTrackByIdDirective, CompactItemRowComponent, NzDividerModule, NgTemplateOutlet, FlexModule, NzButtonModule, NzIconModule, NzToolTipModule, NzWaveModule, MapPositionComponent, NzProgressModule, NzPopconfirmModule, ClipboardDirective, NzSpinModule, AsyncPipe, DecimalPipe, I18nPipe, TranslateModule]
 })
 export class ListDetailsPanelComponent implements OnChanges, OnInit {
 
@@ -234,7 +233,7 @@ export class ListDetailsPanelComponent implements OnChanges, OnInit {
       return;
     }
     if (this.displayRow.zoneBreakdown) {
-      this.zoneBreakdown = new ZoneBreakdown(this.displayRow.rows, this.displayRow.filterChain, this.getHideZoneDuplicates(), this.finalItems);
+      this.zoneBreakdown = new ZoneBreakdown(this.displayRow.rows, this.settings, this.displayRow.filterChain, this.getHideZoneDuplicates(), this.finalItems);
       safeCombineLatest(this.zoneBreakdown.rows
         .map(row => {
           return this.hasPositionsInRows(row.items, row.zoneId).pipe(
@@ -449,12 +448,17 @@ export class ListDetailsPanelComponent implements OnChanges, OnInit {
     })).pipe(
       map(rowsWithNames => {
         return rowsWithNames.reduce((exportString, { row, itemName }) => {
-          return exportString + `${row.amount}x ${itemName}\n`;
+          return this.appendExportStringWithRow(exportString, row, itemName);
         }, `${this.translate.instant(this.displayRow.title)} :\n`);
       }),
       first()
     );
   };
+
+  appendExportStringWithRow(exportString: string, row: ListRow, itemName: string): string {
+    const remaining = row.amount - (row.done || 0);
+    return (remaining > 0) ? (exportString + `${remaining}x ${itemName}\n`) : exportString;
+  }
 
   trackByItem(index: number, item: ListRow): number {
     return item.id;