+++ libs/data/src/lib/extracts-hash.ts
@@ -1,1 +1,1 @@
-export const extractsHash = `1a00d828a8adff214dd2423c5e830d9905a597f4`;
+export const extractsHash = `4abb15ee58ff97d8a8759107bb370ecf51aff872`;